Parsed HID devices
There is an initial HID report descriptor parser in the kernel now.
It's very simple & only handles gamepads (kb & mouse still supported via
boot protocol)

That means you should be able to have some functionnality out of a
gamepad without having it supported in the kernel ! (if planets are
fully aligned).

There might be bugs, but you can see my gamepad is not in the tables
anymore, since it's automatically parsed (incl analog, hat & buttons !)

There might need a buttons remapping app to be more useable (no gamepad
works the same) but that's a start !

Please report non or weird functioning gamepads - note that we
use the parsed descriptor only if it's not in the table, so remove it
if needed (I only removed mine)

Note that dualshock are somewhat special beasts and need separate
updates to fix.

yay !