Make sound run on its own DMA
from now on, sound will run at its own DMA and will not be shared with
Video DAC. That means that sampling frequency is not linked anymore to
vsync, (and will not change with it !) and can be anywhere from 1kHz to
whatever the DAC can produce.

Also, you can block in audio callback and not perturb video out (but
don't).

Tested on real hardware and with hwchip and test_sampler (will need to
update a few other progs), generally nothing to do besides removing
audio_on which does nothing now.