Emulator: fix detection of directories.
d_type causes more trouble than it solves…

1) It is not part of POSIX standards: that only specifies an inode and a
name
(http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/dirent.h.html).
Anything else is platform specific.
2) Even on Linux where this comes from, it is not supported by all
filesystems. The reason for this is that the directory does not
necessarily knows the type of its children (it depends how things are
laid out in the filesystem data structures). So, Linux can return
DT_UNKNOWN and then you have to use stat() anyway.

Implement the simple and sane solution: always use stat() which is the
only proper way to know if an inode is a directory.